server/requirements-api.txt
tinydb==4.7.1
fastapi==0.63.0
starlette==0.13.6
python-multipart==0.0.6
uvicorn[standard]==0.13.4
python-dotenv==1.0.0
